+++ programs/drift/src/error.rs
@@ -625,6 +625,10 @@ pub enum ErrorCode {
     InvalidLiquidateSpotWithSwap,
     #[msg("User in swift message does not match user in ix context")]
     SwiftUserContextUserMismatch,
+    #[msg("User fuel overflow threshold not met")]
+    UserFuelOverflowThresholdNotMet,
+    #[msg("FuelOverflow account not found")]
+    FuelOverflowAccountNotFound,
 }
 
 #[macro_export]