+++ sdk/src/addresses/pda.ts
@@ -72,6 +72,19 @@ export function getUserStatsAccountPublicKey(
 	)[0];
 }
 
+export function getFuelOverflowAccountPublicKey(
+	programId: PublicKey,
+	authority: PublicKey
+): PublicKey {
+	return PublicKey.findProgramAddressSync(
+		[
+			Buffer.from(anchor.utils.bytes.utf8.encode('fuel_overflow')),
+			authority.toBuffer(),
+		],
+		programId
+	)[0];
+}
+
 export function getRFQUserAccountPublicKey(
 	programId: PublicKey,
 	userAccountPublicKey: PublicKey